requirements.txt
pandas
numpy
scipy
statsmodels
scikit-learn
xgboost
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pyarrow
openai
streamlit
plotly
matplotlib
jupyter
pytest
tqdm
google-api-python-client

